refactor: add presets like from markdown-it
