+++ src/extensions/behavior/index.ts
@@ -1,6 +1,5 @@
 import type {ExtensionAuto} from '../../core';
 import {PlaceholderOptions} from '../../utils/placeholder';
-import {Emoji, EmojiOptions} from '../yfm/Emoji';
 
 import {ClicksOnEdges} from './ClicksOnEdges';
 import {Clipboard, ClipboardOptions} from './Clipboard';
@@ -14,12 +13,17 @@ import {Selection} from './Selection';
 import {SelectionContext, SelectionContextOptions} from './SelectionContext';
 import {WidgetDecoration} from './WidgetDecoration';
 
-export * from './Cursor';
-export * from './History';
+export * from './ClicksOnEdges';
 export * from './Clipboard';
-export * from './Selection';
+export * from './CommandMenu';
+export * from './Cursor';
+export * from './FilePaste';
+export * from './History';
 export * from './Placeholder';
 export * from './ReactRenderer';
+export * from './Selection';
+export * from './SelectionContext';
+export * from './WidgetDecoration';
 
 export type BehaviorPresetOptions = {
     cursor?: CursorOptions;
@@ -30,8 +34,6 @@ export type BehaviorPresetOptions = {
     selectionContext?: SelectionContextOptions;
 
     commandMenu?: CommandMenuOptions;
-
-    emoji?: EmojiOptions;
 };
 
 export const BehaviorPreset: ExtensionAuto<BehaviorPresetOptions> = (builder, opts) => {
@@ -44,10 +46,6 @@ export const BehaviorPreset: ExtensionAuto<BehaviorPresetOptions> = (builder, op
         .use(ReactRendererExtension, opts.reactRenderer)
         .use(WidgetDecoration)
         .use(SelectionContext, opts.selectionContext ?? {});
-
-    if (opts.emoji) {
-        builder.use(Emoji, opts.emoji);
-    }
 
     if (opts.commandMenu) builder.use(CommandMenu, opts.commandMenu);
     builder.use(FilePaste);
